Inject initial pull requests
I also moved the PullRequest interface up into the app layer.
