+++ app/javascript/packs/artsy_pull_requests.tsx
@@ -19,6 +19,7 @@ document.addEventListener("DOMContentLoaded", () => {
   }
 
   const props: ArtsyPullRequestsProps = {
+    pullRequests: [],
     startListening,
     stopListening
   }
